refactor(sports-search): move combined-search hooks to own file
LIIKUNTA-340
